quick refactor of DB utils
