+++ lib/db/scraper_data.js
@@ -1,39 +1,16 @@
-const dotenv = require("dotenv");
-dotenv.config();
-const { generateKey } = require("../data/dataDefaulter");
+const {
+    checkItemExistsByRefId,
+    checkItemsExistByRefIds,
+    debugLog,
+    faunaQuery,
+    generateId,
+} = require("./utils");
+const { generateKey } = require("../../data/dataDefaulter");
+const { getGeocode } = require("./../../getGeocode");
 const faunadb = require("faunadb"),
     fq = faunadb.query;
 const lodash = require("lodash");
-const { getGeocode } = require("./../getGeocode");
 
-DEBUG = false; // set to false to disable debugging
-function debug_log(...args) {
-    if (DEBUG) {
-        console.log(...args);
-    }
-}
-
-const client = new faunadb.Client({ secret: process.env.FAUNA_DB });
-
-/* 
-    General utility methods, used just within db-utils and test file.
-*/
-async function faunaQuery(query) {
-    try {
-        debug_log(`trying to execute query ${JSON.stringify(query)}`);
-        const res = await client.query(query);
-        debug_log(
-            `successfully executed query ${JSON.stringify(
-                query
-            )}, got res ${JSON.stringify(res)}`
-        );
-        return res;
-    } catch (error) {
-        console.error(`for query ${JSON.stringify(query)}, got error ${error}`);
-        console.error(error.description);
-        throw error;
-    }
-}
 function hashCode(s) {
     // Taken from: https://gist.github.com/hyamamoto/fd435505d29ebfa3d9716fd2be8d42f0
     var h = 0,
@@ -41,10 +18,6 @@ function hashCode(s) {
         i = 0;
     if (l > 0) while (i < l) h = ((h << 5) - h + s.charCodeAt(i++)) | 0;
     return h;
-}
-
-async function generateId() {
-    return faunaQuery(fq.NewId());
 }
 
 function generateLocationId({ name, street, city, zip }) {
@@ -129,60 +102,6 @@ function addAppointmentsToLocations(locations) {
         }
         return [...(await acc)];
     }, []);
-}
-
-/*
-    Basic CRUD operations.
-*/
-async function retrieveItemByRefId(collectionName, refId) {
-    const result = await faunaQuery(
-        fq.Get(fq.Ref(fq.Collection(collectionName), refId))
-    );
-    debug_log(
-        `querying ${collectionName} collection with refId ${refId} and got result ${JSON.stringify(
-            result
-        )}`
-    );
-    return result;
-}
-
-async function retrieveItemsByRefIds(collectionName, refIds) {
-    const queries = refIds.map((refId) =>
-        fq.Get(fq.Ref(fq.Collection(collectionName), refId))
-    );
-    const result = await faunaQuery(queries);
-    debug_log(
-        `querying ${collectionName} collection with refIds ${refIds} and got result ${JSON.stringify(
-            result
-        )}`
-    );
-    return result;
-}
-
-async function checkItemExistsByRefId(collectionName, refId) {
-    const result = await faunaQuery(
-        fq.Exists(fq.Ref(fq.Collection(collectionName), refId))
-    );
-    return result;
-}
-
-async function checkItemsExistByRefIds(collectionName, refIds) {
-    const queries = refIds.map((refId) =>
-        fq.Exists(fq.Ref(fq.Collection(collectionName), refId))
-    );
-    const result = await faunaQuery(queries);
-    return result;
-}
-
-async function deleteItemByRefId(collectionName, refId) {
-    await faunaQuery(fq.Delete(fq.Ref(fq.Collection(collectionName), refId)));
-}
-
-async function deleteItemsByRefIds(collectionName, refIds) {
-    const queries = refIds.map((refId) =>
-        fq.Delete(fq.Ref(fq.Collection(collectionName), refId))
-    );
-    await faunaQuery(queries);
 }
 
 async function writeLocationByRefId({
@@ -309,7 +228,7 @@ async function writeAppointmentsByRefIds(locationsWithAllData) {
             )
         )
         .flat();
-    debug_log("writeAppointmentsByRefIds queries", JSON.stringify(queries));
+    debugLog("writeAppointmentsByRefIds queries", JSON.stringify(queries));
     const result = await faunaQuery(queries);
     return result;
 }
@@ -329,7 +248,7 @@ async function getScraperRunsByLocation(locationRefId) {
             fq.Lambda((x) => fq.Get(x))
         )
     );
-    debug_log(
+    debugLog(
         `for locationRefId ${locationRefId} got response ${JSON.stringify(
             scraperRuns
         )}`
@@ -365,7 +284,7 @@ async function getAppointmentsByScraperRun(scraperRunRefId) {
             fq.Lambda((x) => fq.Get(x))
         )
     );
-    debug_log(
+    debugLog(
         `for scraperRunRefId ${scraperRunRefId} got response ${JSON.stringify()}`
     );
     return appointments;
@@ -385,7 +304,7 @@ async function getAppointmentsByScraperRuns(scraperRunRefIds) {
     );
     const appointments = await faunaQuery(queries);
 
-    debug_log(
+    debugLog(
         `for scraperRunRefIds ${scraperRunRefIds} got response ${JSON.stringify(
             appointments
         )}`
@@ -413,7 +332,7 @@ async function writeScrapedData({
 }) {
     const locationRefId = generateLocationId({ name, street, city, zip });
     const itemExists = await checkItemExistsByRefId("locations", locationRefId);
-    debug_log(`item ${locationRefId} exists: ${itemExists}`);
+    debugLog(`item ${locationRefId} exists: ${itemExists}`);
     if (!itemExists) {
         const locationData = await getGeocode(name, street, zip);
         await writeLocationByRefId({
@@ -498,7 +417,7 @@ async function writeScrapedDataBatch(locations, timestamp) {
     const locationsWithAllData = await addAppointmentsToLocations(
         locationsWithScraperRunIds
     );
-    debug_log("locationsWithAllData", locationsWithAllData);
+    debugLog("locationsWithAllData", locationsWithAllData);
     writeAppointmentsByRefIds(locationsWithAllData);
 }
 
@@ -507,21 +426,14 @@ async function writeScrapedDataBatch(locations, timestamp) {
     This will go into a lambda.
 */
 async function getAppointmentsForAllLocations() {}
-
 module.exports = {
     addGeneratedIdsToLocations,
-    checkItemExistsByRefId,
-    checkItemsExistByRefIds,
     getAppointmentsForAllLocations,
     getAppointmentsByScraperRun,
     getAppointmentsByScraperRuns,
-    deleteItemByRefId,
-    deleteItemsByRefIds,
     generateLocationId,
     getScraperRunsByLocation,
     getScraperRunsByLocations,
-    retrieveItemByRefId,
-    retrieveItemsByRefIds,
     writeAppointmentsByRefId,
     writeAppointmentsByRefIds,
     writeLocationByRefId,
